requirements.txt
fastapi
supabase
cerebras-cloud-sdk
uvicorn
